+++ src/app/[locale]/events/[state]/[eventSlug]/page.tsx
@@ -4,7 +4,7 @@ import { permanentRedirect } from 'next/navigation'
 import { EventDialogContent } from '@/components/app/pageEvents/components/eventDialogContent'
 import { EventsPageDialogDeeplinkLayout } from '@/components/app/pageEvents/eventsPageDialogDeeplinkLayout'
 import { PageProps } from '@/types'
-import { getEvent } from '@/utils/server/builderIO/swc-event'
+import { getEvent } from '@/utils/server/builderIO/swcEvent'
 import { generateMetadataDetails } from '@/utils/server/metadataUtils'
 import { getIntlUrls } from '@/utils/shared/urls'
 import { US_MAIN_STATE_CODE_TO_DISPLAY_NAME_MAP } from '@/utils/shared/usStateUtils'
